Change Salon Manager Area, add View Component to get list of salons

+++ Services/BeautyBooking.Services.Data/Salons/ISalonsService.cs
@@ -17,6 +17,6 @@
 
         Task DeleteAsync(string id);
 
-        Task RateSalon(string id, double rateValue);
+        Task RateSalon(string id, int rateValue);
     }
 }
